modify code structure
