+++ mistralrs-core/src/vision_models/llava/mod.rs
@@ -1,5 +1,6 @@
 pub mod llava_llm;
 pub mod llava_next;
 pub mod llava_next_inputs_processor;
+mod utils;
 //pub mod llava;
 //pub mod llava_inputs_processor;
